feat(lazy_hgraph): support extra info filtering

+++ src/algorithm/hgraph/hgraph_parameter.cpp
@@ -149,7 +149,7 @@ HGraphParameter::ToJson() const {
     json[REORDER_SOURCE_KEY].SetString(this->reorder_source);
     json[BASE_CODES_KEY].SetJson(this->base_codes_param->ToJson());
     json[GRAPH_KEY].SetJson(this->bottom_graph_param->ToJson());
-    json[EF_CONSTRUCTION_KEY].SetInt(this->ef_construction);
+    json[EF_CONSTRUCTION_KEY].SetUint64(this->ef_construction);
     json[ALPHA_KEY].SetFloat(this->alpha);
     json[SUPPORT_DUPLICATE].SetBool(this->support_duplicate);
     json[DUPLICATE_DISTANCE_THRESHOLD].SetFloat(this->duplicate_distance_threshold);
