+++ src/datacell/compressed_graph_datacell_parameter.h
@@ -44,7 +44,7 @@ public:
     JsonType
     ToJson() const override {
         JsonType json;
-        json[GRAPH_PARAM_MAX_DEGREE_KEY].SetInt(this->max_degree_);
+        json[GRAPH_PARAM_MAX_DEGREE_KEY].SetUint64(this->max_degree_);
         json[GRAPH_STORAGE_TYPE_KEY].SetString(GRAPH_STORAGE_TYPE_VALUE_COMPRESSED);
         json[SUPPORT_DUPLICATE].SetBool(this->support_duplicate_);
         json[HGRAPH_USE_REVERSE_EDGES_KEY].SetBool(this->use_reverse_edges_);
